+++ src/main/java/com/nuriza/fqw/fqw/services/impl/ClientServiceImpl.java
@@ -28,7 +28,7 @@ public class ClientServiceImpl implements ClientService {
     }
 
     @Override
-    public Client getById(Integer id) {
+    public Client getById(Long id) {
         return repository.findById(id).orElse(new Client());
     }
